fix: keep deps surface out of commands
