+++ src/commands/json_output/ops.rs
@@ -2,7 +2,7 @@ use crate::cli_surface::Commands;
 
 use super::{map, JsonRun};
 use crate::commands::{
-    api, auth, ci, daemon, db, deploy, deps, file, git, http, issues, logs, self_cmd, server, ssh,
+    api, auth, ci, daemon, db, deploy, file, git, http, issues, logs, self_cmd, server, ssh,
     status, triage, upgrade, GlobalArgs,
 };
 
@@ -13,7 +13,7 @@ pub(super) fn dispatch(command: Commands, global: &GlobalArgs) -> JsonRun {
         Commands::Ssh(args) => map(ssh::run(args, global)),
         Commands::Server(args) => map(server::run(args, global)),
         Commands::Db(args) => map(db::run(args, global)),
-        Commands::Deps(args) => map(deps::run(args, global)),
+        Commands::Deps(args) => map(args.run()),
         Commands::File(args) => map(file::run(args, global)),
         Commands::Logs(args) => map(logs::run(args, global)),
         Commands::Triage(args) => map(triage::run(args, global)),
